+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # [ES_Utilities](https://github.com/ESpmdark/ES_Utilities)
 
-## Version: 2.1.6
-- If using Plumber, add its housing macro to the travelframe.
+## Version: N/A
+- Log is now contained in the addon.